+++ src/com/intellij/idea/plugin/hybris/tools/ccv2/dto/CCv2Environment.kt
@@ -31,6 +31,7 @@ data class CCv2Environment(
     var deployedBuild: CCv2Build? = null,
     val dynatraceLink: String? = null,
     val loggingLink: String? = null,
+    val problems: Int? = null,
 ) : CCv2DTO, Comparable<CCv2Environment> {
 
     override fun compareTo(other: CCv2Environment) = name.compareTo(other.name)