+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Do not retrieve build for environments with `UNDEPLOYED` and `UNKNOWN` deployment status [#1101](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1101)
 - Integrated with CCv2 API V1 to show Dynatrace link [#1102](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1102)
 - Show OpenSearch link for each environment [#1103](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1103)
+- Show the number of problems for each environment [#1104](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1104)
 
 ### `Project Import` enhancements
 - Loading backoffice sources provided by SAP on project import [#1096](https://github.com/epam/sap-commerce-intellij-idea-plugin/issues/1096)